+++ resources/js/components/AlbumCard.vue
@@ -40,6 +40,7 @@
 
                 <div class="dropdown-menu dropdown-menu-right" aria-labelledby="dropdownMenuButton">
                     <a class="dropdown-item" :href="album.permalink"><i class="fas fa-link fa-fw mr-1"></i> {{ __('Permalink') }}</a>
+                    <a class="dropdown-item" :href="album.url_download" v-if="album.url_download"><i class="fas fa-file-download fa-fw mr-1"></i> {{ __('Download') }}</a>
 
                     <h6 class="dropdown-header" v-if="album.editable">{{ __('Manage') }}</h6>